# filefn/server/contracts/filefn-client-v1.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FileFn Client Contract",
    "version": "v1",
    "description": "Deterministic OpenAPI contract for the FileFn HTTP surface supported by the Swift client."
  },
...
  "servers": [
    {
      "url": "/filefn",
      "description": "Canonical FileFn base path"
    }
  ],
...
  "paths": {
...
    "/upload/init": {
      "post": {
        "summary": "Initialize an upload session",
        "tags": [
          "uploads"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestID"
          },
          {
            "$ref": "#/components/parameters/XFileFnClientVersion"
          },
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/XIdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUploadSessionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UploadSession"
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Canonical FileFn error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/policies": {
      "get": {
        "summary": "List policies",
        "tags": [
          "policies"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestID"
          },
          {
            "$ref": "#/components/parameters/XFileFnClientVersion"
          },
          {
            "$ref": "#/components/parameters/Authorization"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/PolicyPage"
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Canonical FileFn error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/quota/storage": {
      "get": {
        "summary": "Get storage quota",
        "tags": [
          "quota"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/XRequestID"
          },
          {
            "$ref": "#/components/parameters/XFileFnClientVersion"
          },
          {
            "$ref": "#/components/parameters/Authorization"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessEnvelope"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/StorageQuota"
                        }
                      },
                      "required": [
                        "data"
                      ]
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "Canonical FileFn error envelope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "XRequestID": {
        "name": "x-request-id",
        "in": "header",
        "required": false,
        "description": "Opaque request correlation ID forwarded through FileFn envelopes.",
        "schema": {
          "type": "string"
        }
      },
      "XFileFnClientVersion": {
        "name": "x-filefn-client-version",
        "in": "header",
        "required": false,
        "description": "Optional client version header emitted by FileFn clients.",
        "schema": {
          "type": "string"
        }
      },
      "Authorization": {
        "name": "Authorization",
        "in": "header",
        "required": false,
        "description": "Bearer or custom authorization header resolved by the host app.",
        "schema": {
          "type": "string"
        }
      },
      "XIdempotencyKey": {
        "name": "x-idempotency-key",
        "in": "header",
        "required": false,
        "description": "Optional idempotency key used for upload initialization.",
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ]
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "const": false
          },
          "requestId": {
            "type": "string"
          },
          "error": {
            "$ref": "#/components/schemas/ServerError"
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
      "ServerError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "JSONValue": {
        "oneOf": [
          {
            "type": "string"
          },
          {
            "type": "number"
          },
          {
            "type": "integer"
          },
          {
            "type": "boolean"
          },
          {
            "type": "object",
            "additionalProperties": true
          },
          {
            "type": "array",
            "items": {}
          },
          {
            "type": "null"
          }
        ]
      },
...
      "CreateUploadSessionRequest": {
        "type": "object",
        "properties": {
          "policy": {
            "type": "string"
          },
          "fileName": {
            "type": "string"
          },
          "size": {
            "type": "integer"
          },
          "mimeType": {
            "type": "string"
          },
          "fileId": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/JSONValue"
            }
          }
        },
        "required": [
          "policy",
          "fileName",
          "size",
          "mimeType"
        ]
      },
      "UploadSession": {
        "type": "object",
        "properties": {
          "uploadSessionId": {
            "type": "string"
          },
          "uploadSessionToken": {
            "type": "string"
          },
          "uploadMode": {
            "type": "string",
            "enum": [
              "multipart-signed-url",
              "proxy"
            ]
          },
          "chunkSizeBytes": {
            "type": "integer"
          },
          "totalParts": {
            "type": "integer"
          },
          "expiresAt": {
            "type": "string"
          }
        },
        "required": [
          "uploadSessionId",
          "uploadMode",
          "chunkSizeBytes",
          "totalParts",
          "expiresAt"
        ]
      },
...
      "PolicySummary": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "maxSizeBytes": {
            "type": "integer"
          },
          "contentTypes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "visibility": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "maxSizeBytes",
          "contentTypes",
          "visibility"
        ]
      },
      "PolicyPage": {
        "type": "object",
        "properties": {
          "policies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PolicySummary"
            }
          }
        },
        "required": [
          "policies"
        ]
      },
      "StorageQuota": {
        "type": "object",
        "properties": {
          "current": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        },
        "required": [
          "current",
          "limit"
        ]
      },
...
    }
  }
}